+++ config/loader.js
@@ -14,8 +14,8 @@ exports.babel = {
 exports.eslint = {
   enable: true,
   test: /\.jsx?$/,
-  loader: 'eslint-loader',
-  exclude: [/node_modules/],
+  use: ['eslint-loader'],
+  exclude: /node_modules/,
   enforce: 'pre'
 };