+++ test/loader.test.js
@@ -4,17 +4,32 @@ const WebpackTool = require('webpack-tool');
 const webpack = WebpackTool.webpack;
 const merge = WebpackTool.merge;
 const WebpackBaseBuilder = require('../lib/base');
-const Loader = require('../utils/loader');
 const path = require('path').posix;
+
 // http://chaijs.com/api/bdd/
-function createBuilder() {
-  const builder = new WebpackBaseBuilder();
-  builder.setBuildPath(path.join(__dirname, 'test'));
+function createBuilder(config) {
+  const builder = new WebpackBaseBuilder(config);
+  builder.setBuildPath(path.join(__dirname, 'dist/loader'));
   builder.setPublicPath('/public');
   builder.setEntry({
     include: path.join(__dirname, 'test')
   });
   return builder;
+}
+
+function getLoaderByName(name, rules) {
+  const loaderName = `${name}-loader`;
+  return rules.find(rule => {
+    return rule.use.some(loader => {
+      return loaderName === loader || (typeof loader === 'object' && loader.loader === loaderName);
+    });
+  });
+}
+
+function getLoaderByTest(test, rules) {
+  return rules.find(rule => {
+    return rule.test.toString() === test.toString();
+  });
 }
 
 describe('loader.test.js', () => {
@@ -30,137 +45,94 @@ describe('loader.test.js', () => {
   afterEach(() => {
   });
 
-  describe('#webpack loader update test', () => {
-    it('should updateLoader', () => {
-      const builder = createBuilder();
+  describe('#webpack createWebpackLoader test', () => {
+    it('should loader enable test', () => {
+      const builder1 = createBuilder();
+      const webpackConfig1 = builder1.create();
+      const rules1 = webpackConfig1.module.rules;
+      expect(getLoaderByName('eslint', rules1)).to.include.all.keys(['test', 'use']);
+      expect(getLoaderByName('babel', rules1)).to.include.all.keys(['test', 'use']);
+      expect(getLoaderByTest(/\.(woff2?|eot|ttf|otf)(\?.*)?$/, rules1)).to.include.all.keys(['test', 'use']);
 
-      builder.setStyleLoaderName('vue-style-loader');
-      builder.addLoader(/\.vue$/, 'vue-loader', () => ({
-        options: Loader.getStyleLoaderOption(builder.getStyleConfig())
-      }));
-      builder.updateLoader({
-        test: /\.vue$/,
-        loader: 'vue-loader',
-        options: {
-          compilerModules: [{
-            postTransformNode: el => {
-              el.staticStyle = `$processStyle(${el.staticStyle})`;
-              el.styleBinding = `$processStyle(${el.styleBinding})`;
-            }
-          }]
+      const builder2 = createBuilder({
+        loaders: {
+          eslint: false,
+          babel: false,
+          urlfont: false
         }
       });
 
-      const newLoaderIndex = builder.findLoaderIndex('vue-loader', 'loader');
+      const webpackConfig2 = builder2.create();
+      const rules2 = webpackConfig2.module.rules;
+      expect(getLoaderByName('eslint', rules2)).to.equal(undefined);
+      expect(getLoaderByName('babel', rules2)).to.be.undefined;
+      expect(getLoaderByName(/\.(woff2?|eot|ttf|otf)(\?.*)?$/, rules2)).to.be.undefined;
+    });
 
-      const newLoader = builder.configLoader[newLoaderIndex];
-
-      expect(typeof newLoader.fn === 'function').to.be.true;
-      expect(newLoader.options).to.have.property('compilerModules');
+    it('should loader merge options test', () => {
+      const config = {
+        loaders: {
+          eslint: {
+            options: {
+              fix: true
+            }
+          },
+          babel: {
+            options: {
+              presets: ['es2015', 'stage-2'],
+              plugins: [
+                'add-module-exports'
+              ],
+              comments: false
+            }
+          }
+        }
+      };
+      const builder = createBuilder(config);
 
       const webpackConfig = builder.create();
       const rules = webpackConfig.module.rules;
-      const vueLoaderIndex = rules.findIndex(rule => rule.loader.includes('vue-loader'));
-      const vueLoader = rules[vueLoaderIndex];
-      expect(vueLoader.options).to.have.property('loaders');
-      expect(vueLoader.options).to.have.property('compilerModules');
+      const eslint = getLoaderByName('eslint', rules);
+      const babel = getLoaderByName('babel', rules);
+      expect(eslint.use[0].options.fix).to.be.true;
+      expect(babel.use[0].options).to.include.all.keys(['presets', 'plugins']);
+      expect(babel.use[0].options.comments).to.be.false;
     });
-  });
 
-  describe('#webpack loader delete test', () => {
-    const imageLoader = {
-      test: /\.(png|jpe?g|gif|svg)(\?.*)?$/,
-      loader: 'url-loader',
-      query: {
-        limit: 2048
-      }
-    };
-    const builder = createBuilder();
-    expect(builder.findLoaderIndex(imageLoader) > -1);
-
-    builder.deleteLoader(imageLoader);
-    expect(builder.findLoaderIndex(imageLoader) === -1);
-
-    builder.addLoader(imageLoader);
-    expect(builder.findLoaderIndex(imageLoader) > -1);
-    builder.deleteLoader('url-loader');
-    expect(builder.findLoaderIndex(imageLoader) === -1);
-
-    builder.addLoader(imageLoader);
-    expect(builder.findLoaderIndex(imageLoader) > -1);
-    builder.deleteLoader({loader: 'url-loader'}, 'loader');
-    expect(builder.findLoaderIndex(imageLoader) === -1);
-
-    builder.addLoader(imageLoader);
-    expect(builder.findLoaderIndex(imageLoader) > -1);
-    builder.deleteLoader({test: /\.(png|jpe?g|gif|svg)(\?.*)?$/}, 'test');
-    expect(builder.findLoaderIndex(imageLoader) === -1);
-
-    expect(builder.deleteLoader({test: /\.(png|jpe?g|gif|svg)(\?.*)?$/}) === null );
-  });
-
-  describe('#webpack loader find test', () => {
-
-  });
-
-  describe('#webpack loader add test', () => {
-
-    const imageLoader = {
-      test: /\.(png|jpe?g|gif|svg)(\?.*)?$/,
-      loader: 'url-loader',
-      query: {
-        limit: 2048
-      }
-    };
-
-    const fontLoader = {
-      test: /\.(woff2?|eot|ttf|otf)(\?.*)?$/,
-      loader: 'url-loader',
-      query: {
-        limit: 2048
-      }
-    };
-
-    it('should addLoader#test_loader_replace_and_update', () => {
-      const builder = createBuilder();
-      builder.setConfig({
-        loaderOption: {
-          imageUrl: { query: { limit: 2048 } }
+    it('should loader use override test', () => {
+      const config = {
+        loaders: {
+          babel: {
+            use: ['babel-loader', 'eslint-loader']
+          },
+          options: {
+            babel: {
+              options:{
+                presets: ['es2015', 'stage-2'],
+                plugins: [
+                  'add-module-exports'
+                ],
+                comments: false
+              }
+            },
+            eslint: {
+              options:{
+                fix: true
+              }
+            }
+          }
         }
+      };
+      const builder = createBuilder(config);
+      const webpackConfig = builder.create();
+      const rules = webpackConfig.module.rules;
+      const loader = rules.find(rule => {
+        return rule.use.length === 2 && rule.use[0].loader === 'babel-loader' && rule.use[1].loader === 'eslint-loader'
       });
-      const loaderIndex = builder.findLoaderIndex(imageLoader);
-      expect(loaderIndex > -1);
-      expect(builder.configLoader[loaderIndex].fn().query.limit).to.equal(2048);
-    });
+      expect(!!loader).to.be.true;
+      expect(loader.use[0].options).to.include.all.keys(['presets', 'plugins']);
+      expect(loader.use[1].options.fix).to.be.true;
 
-    it('should addLoader#test_same_loader_name_add', () => {
-      const builder = createBuilder();
-      builder.addLoader(imageLoader);
-      builder.addLoader(fontLoader);
-      const imageLoaderIndex = builder.findLoaderIndex(imageLoader, 'all');
-      const fontLoaderIndex = builder.findLoaderIndex(fontLoader, 'all');
-
-      expect(imageLoaderIndex > -1);
-      expect(builder.configLoader[imageLoaderIndex].test.toString()).to.equal(imageLoader.test.toString());
-
-      expect(fontLoaderIndex > -1);
-      expect(builder.configLoader[fontLoaderIndex].test.toString()).to.equal(fontLoader.test.toString());
-    });
-
-    it('should addLoader#test_loader_option_function', () => {
-      const builder = createBuilder();
-      builder.addLoader(fontLoader, null, null, null, 'replace');
-      builder.addLoader(fontLoader.test, fontLoader.loader, () => merge({ query: fontLoader.query }, { query: { name: 'font-url-loader' } }), null, 'replace');
-
-      const urlLoaders = () => builder.configLoader.filter(item => item.test.toString() === fontLoader.test.toString());
-      expect(urlLoaders().length).to.equal(1);
-
-      const fontLoaderIndex = builder.findLoaderIndex(fontLoader, 'all');
-      const expectLoader = builder.configLoader[fontLoaderIndex];
-
-      expect(expectLoader).to.have.property('fn');
-      expect(expectLoader.fn().query.limit).to.equal(2048);
-      expect(expectLoader.fn().query.name).to.equal('font-url-loader');
     });
   });
 });